# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sensor Management API Documentation",
        "description": "API Documentation for Sensor Management System",
        "contact": {
            "name": "",
            "email": ""
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/settings": {
            "get": {
                "tags": [
                    "Settings"
                ],
                "summary": "Get user notification settings",
                "description": "Returns the current user's notification settings",
                "operationId": "getUserSettings",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SettingResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Settings"
                ],
                "summary": "Update user notification settings",
                "description": "Updates the current user's notification settings",
                "operationId": "updateUserSettings",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateSettingRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "type": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Settings updated successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - Email not verified"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UpdateSettingRequest": {
                "required": [
                    "settings"
                ],
                "properties": {
                    "settings": {
                        "type": "array",
                        "items": {
                            "required": [
                                "id",
                                "email_notification_allowed"
                            ],
                            "properties": {
                                "id": {
                                    "description": "Setting ID",
                                    "type": "integer"
                                },
                                "email_notification_allowed": {
                                    "description": "Whether email notifications are enabled",
                                    "type": "boolean"
                                },
                                "threshold": {
                                    "description": "Notification threshold (required for non-binary sensors when notifications are enabled)",
                                    "type": "number",
                                    "format": "float",
                                    "nullable": true
                                },
                                "cooldown": {
                                    "description": "Notification cooldown period in minutes (required when notifications are enabled)",
                                    "type": "integer",
                                    "nullable": true
                                },
                                "min_unit_diff": {
                                    "description": "Minimum unit difference to trigger notification",
                                    "type": "number",
                                    "format": "float",
                                    "nullable": true
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "EnumResource": {
                "properties": {
                    "value": {
                        "description": "Enum value",
                        "type": "string"
                    },
                    "label": {
                        "description": "Human-readable label",
                        "type": "string"
                    },
                    "symbol": {
                        "description": "Symbol representation",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "SensorResource": {
                "properties": {
                    "id": {
                        "description": "Sensor ID",
                        "type": "integer"
                    },
                    "sensor_name": {
                        "description": "Name of the sensor",
                        "type": "string"
                    },
                    "type": {
                        "description": "Type identifier of the sensor",
                        "type": "string"
                    },
                    "display_name": {
                        "description": "Display name of the sensor",
                        "type": "string"
                    },
                    "unit_of_measurement": {
                        "description": "Unit of measurement for sensor readings",
                        "type": "string",
                        "nullable": true
                    },
                    "is_output_binary": {
                        "description": "Whether sensor output is binary",
                        "type": "boolean"
                    },
                    "image_path": {
                        "description": "Path to sensor image",
                        "type": "string",
                        "nullable": true
                    },
                    "icon_path": {
                        "description": "Path to sensor icon",
                        "type": "string",
                        "nullable": true
                    },
                    "color_class": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/EnumResource"
                            }
                        ],
                        "nullable": true
                    },
                    "group_name": {
                        "description": "Name of the sensor group",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "SettingResource": {
                "properties": {
                    "id": {
                        "description": "Setting ID",
                        "type": "integer"
                    },
                    "email_notification_allowed": {
                        "description": "Whether email notifications are enabled",
                        "type": "boolean"
                    },
                    "threshold": {
                        "description": "Notification threshold value",
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "cooldown": {
                        "description": "Notification cooldown period in minutes",
                        "type": "integer",
                        "nullable": true
                    },
                    "min_unit_diff": {
                        "description": "Minimum unit difference to trigger notification",
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "sensor": {
                        "$ref": "#/components/schemas/SensorResource"
                    }
                },
                "type": "object"
            },
...
            "ValidationErrorResponse": {
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "error"
                    },
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            }
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sensorAdminKey": {
                "type": "apiKey",
                "name": "X-Admin-Token",
                "in": "header"
            }
        }
    },
...
}